backend/ivc_system/requirements.txt
SQLAlchemy~=2.0.4
uvicorn~=0.20.0
fastapi~=0.92.0
pydantic~=1.10.5
passlib~=1.7.4
alembic~=1.9.4
python-jose~=3.3.0